test_welcome/requirements.txt
fastapi
uvicorn[standard]
pydantic

sqlalchemy
asyncpg
aiosqlite


python-dotenv

pytest
pytest-asyncio
httpx